Created some more Ui elements and previews
