+++ .idea/misc.xml
@@ -10,6 +10,9 @@
         <entry key="../../../../layout/compose-model-1629502005973.xml" value="0.1366797701844572" />
         <entry key="../../../../layout/compose-model-1629518638634.xml" value="0.25" />
         <entry key="../../../../layout/compose-model-1629521179255.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1629588770505.xml" value="0.9754464285714286" />
+        <entry key="../../../../layout/compose-model-1629594954675.xml" value="0.5250244691873371" />
+        <entry key="../../../../layout/compose-model-1629595659103.xml" value="0.31571941458935077" />
       </map>
     </option>
   </component>
