+++ app/src/main/java/com/deanthonee/composepickpocket/utils/FakeData.kt
@@ -26,9 +26,72 @@ class FakeData {
             "Zappy",
             "Xina",
             "Xeolyne",
-            "Stacy"
+            "Stacy",
+            "Udes",
+            "Urchala",
+            "Sarah",
+            "Stacy",
+            "Lee",
+            "Rock",
+            "Rachel",
+            "Jamie",
+            "Jill",
+            "Jerry",
+            "Rick",
+            "Morty",
+            "Tilly",
+            "Candy",
+            "Celina",
+            "Eddy",
+            "Erin",
+            "Javin",
+            "Junior",
+            "Jerome",
+            "Heather",
+            "Heman",
+            "Hurmon",
+            "Ian",
+            "Illy",
+            "Nina",
+            "Nancy",
+            "Nelly",
+            "Mindy",
+            "Mandy",
+            "Niko",
+            "Ossim",
+            "Queen",
+            "Quila",
+            "Veronica",
+            "Van",
+            "Venison",
+            "Wally",
+            "Wiltrox",
+            "Willow",
+            "Zinny"
         )
         val listOfLastNames = listOf(
+            "Uchia",
+            "Uzamaki",
+            "Keen",
+            "Landings",
+            "Cummings",
+            "Joels",
+            "Kent",
+            "Lane",
+            "Wayne",
+            "Rogers",
+            "Lord",
+            "Gurchin",
+            "Plorgio",
+            "Flores",
+            "Guazena",
+            "Horston",
+            "Hill",
+            "Inges",
+            "Ichiro",
+            "Jameson",
+            "Kelly",
+            "Kourn",
             "Johnson",
             "Stevens",
             "Tillsons",
